Part of the updates to EGTS v2, add EGTS_SR_SERVICE_INFO support

+++ cli/receiver/server/server.go
@@ -59,7 +59,7 @@ func (s *Server) handleConn(conn net.Conn) {
 		serviceType       uint8
 		srResponsesRecord egts.RecordDataSet
 		recvPacket        []byte
-		client            uint32
+		client            uint64
 	)
 
 	if s.store == nil {
